pyproject.toml
[tool.poetry]
name = "cc-tweaked-images"
version = "0.1.0"
description = ""
authors = ["TurquoisePenguin7 <turquoisehairedpenguinhorn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
